Extract e2e deprecated helpers
